+++ s4e-web/src/app/common/store.util.ts
@@ -1,7 +1,7 @@
-import { HttpClient } from '@angular/common/http';
 import { Store } from '@datorama/akita';
-import { catchError, map, finalize, shareReplay } from 'rxjs/operators';
+import { catchError, finalize, shareReplay, map } from 'rxjs/operators';
 import { throwError, OperatorFunction, Observable, of } from 'rxjs';
+import { HttpClient } from '@angular/common/http';
 
 export function httpGetRequest$<T>(http: HttpClient, url: string, store: Store<any>): Observable<T | null> {
   store.setLoading(true);
@@ -17,24 +17,18 @@ export function httpPutRequest$<T>(http: HttpClient, url: string, object: T, sto
     .pipe(handleHttpUpdateRequest(store));
 }
 
+export function httpDeleteRequest$(http: HttpClient, url: string, store: Store<any>) {
+  store.setLoading(true);
+  store.setError(null);
+  return http.delete(url)
+    .pipe(handleHttpUpdateRequest(store));
+}
+
 export function httpPostRequest$<T>(http: HttpClient, url: string, object: T, store: Store<any>) {
   store.setLoading(true);
   store.setError(null);
   return http.post<T>(url, object)
     .pipe(handleHttpUpdateRequest(store));
-}
-
-export function httpDeleteRequest$(http: HttpClient, url: string, store: Store<any>) {
-  store.setLoading(true);
-  store.setError(null);
-  return http.delete(url).pipe(handleHttpUpdateRequest(store));
-}
-
-export function catchErrorAndHandleStore<T>(store: Store<any>): OperatorFunction<T, T> {
-  return catchError(error => {
-      store.setError(error);
-      return throwError(error);
-    });
 }
 
 export function handleHttpUpdateRequest(store: Store<any>) {
@@ -56,4 +50,11 @@ export function handleHttpGetRequest(store: Store<any>) {
         shareReplay(1)
       );
   };
+}
+
+export function catchErrorAndHandleStore<T>(store: Store<any>): OperatorFunction<T, T> {
+  return catchError(error => {
+      store.setError(error);
+      return throwError(error);
+    });
 }