+++ s4e-web/src/app/views/settings/groups/state/group.model.ts
@@ -1,4 +1,5 @@
 export interface Group {
+  id: string;
   slug: string;
   name: string;
 }