+++ s4e-web/src/app/views/settings/people/person-list/person-list.component.spec.ts
@@ -1,13 +1,23 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import { InstitutionService } from './../../state/institution/institution.service';
+import { PersonService } from './../state/person.service';
+import { By } from '@angular/platform-browser';
+import { of } from 'rxjs';
+import { PersonQuery } from './../state/person.query';
+import { async, ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {TestingConfigProvider} from '../../../../app.configuration.spec';
 import {PersonListComponent} from './person-list.component';
 import {SettingsModule} from '../../settings.module';
+import { DebugElement } from '@angular/core';
 
 describe('PeopleComponent', () => {
   let component: PersonListComponent;
   let fixture: ComponentFixture<PersonListComponent>;
+  let personQuery: PersonQuery;
+  let institutionService: InstitutionService;
+  let personService: PersonService;
+  let de: DebugElement;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -16,14 +26,22 @@ describe('PeopleComponent', () => {
         RouterTestingModule,
         HttpClientTestingModule
       ],
-      providers: [TestingConfigProvider]
+      providers: [
+        TestingConfigProvider
+      ]
     })
       .compileComponents();
+
+    personQuery = TestBed.get(PersonQuery);
+    personService = TestBed.get(PersonService);
+    institutionService = TestBed.get(InstitutionService);
   }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(PersonListComponent);
     component = fixture.componentInstance;
+    de = fixture.debugElement;
+
     fixture.detectChanges();
   });