+++ s4e-web/src/app/views/settings/people/state/person.service.ts
@@ -1,11 +1,11 @@
+import { httpGetRequest$, httpPostRequest$, httpPutRequest$ } from 'src/app/common/store.util';
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {PersonStore} from './person.store';
 import {S4eConfig} from '../../../../utils/initializer/config.service';
 import {Observable} from 'rxjs';
-import {finalize} from 'rxjs/operators';
 import {DEFAULT_GROUP_SLUG, Person, PersonForm} from './person.model';
-import {catchErrorAndHandleStore, httpPostRequest$} from '../../../../common/store.util';
+import {catchErrorAndHandleStore} from '../../../../common/store.util';
 import {InstitutionQuery} from '../../state/institution/institution.query';
 
 @Injectable({providedIn: 'root'})
@@ -17,26 +17,23 @@ export class PersonService {
   }
 
   fetchAll(institutionSlug: string) {
-    this.store.setLoading(true);
-    this.store.setError(null);
-    this.http.get<Person[]>(`${this.config.apiPrefixV1}/institutions/${institutionSlug}/groups/${DEFAULT_GROUP_SLUG}/members`)
-      .pipe(
-        finalize(() => this.store.setLoading(false)),
-        catchErrorAndHandleStore(this.store)
-      )
+    const url = `${this.config.apiPrefixV1}/institutions/${institutionSlug}/groups/${DEFAULT_GROUP_SLUG}/members`;
+    httpGetRequest$<Person[]>(this.http, url, this.store)
       .subscribe(data => this.store.set(data));
   }
 
-  create$(institutionSlug: string, value: PersonForm): Observable<Person> {
-    this.store.setLoading(true);
+  create$(institutionSlug: string, value: PersonForm): Observable<boolean> {
+    const groupSlugs = [...value.groupSlugs._, 'default'];
+    const person = {...value, groupSlugs} as undefined as Person;
+    const url = `${this.config.apiPrefixV1}/institutions/${institutionSlug}/users`;
+    return httpPostRequest$(this.http, url, person, this.store);
+  }
 
-    return this.http.post<Person>(
-      `${this.config.apiPrefixV1}/institutions/${institutionSlug}/users`,
-      {...value, groupSlugs: [...value.groupSlugs._, 'default']}
-    ).pipe(
-      catchErrorAndHandleStore(this.store),
-      finalize(() => this.store.setLoading(false))
-    );
+  update$(institutionSlug: string, value: PersonForm) {
+    const groupSlugs = [...value.groupSlugs._, 'default'];
+    const person = {...value, groupSlugs} as undefined as Person;
+    const url = `${this.config.apiPrefixV1}/institutions/${institutionSlug}/users`;
+    return httpPutRequest$(this.http, url, person, this.store);
   }
 
   deleteMember(userId: string) {